Address comments from part 1
